+++ app/Http/Controllers/Dashboard/SeriesController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Dashboard;
 
+use App\Repositories\SeriesListsRepository;
 use App\Repositories\SeriesRepository;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -34,7 +35,7 @@ class SeriesController extends Controller
     {
         $condition = array(['id', '>', 0]);
         $series = $this->seriesRepository->findAllBy($condition);
-        $series = $series->sortByDesc('id');
+//        $series = $series->sortByDesc('id');
 
         $siteVar = $this->siteVar;
         $loginUser = $this->loginUser;
@@ -62,15 +63,20 @@ class SeriesController extends Controller
         return redirect(URL_DASHBOARD_SERIES);
     }
 
-    public function doDelete(Request $request)
+    public function doDelete(Request $request, SeriesListsRepository $seriesListsRepository)
     {
         $seriesId = $request->get('series_id');
-        $bool = $this->seriesRepository->delete(['id' => $seriesId]);
 
+        $seriesLists = $seriesListsRepository->findAllBy(['series_id' => $seriesId]);
+        if ($seriesLists->count() > 0) {
+            return response()->json(['status' => false, 'message' => '目前有產品屬於該系列，無法刪除']);
+        }
+
+        $bool = $this->seriesRepository->delete(['id' => $seriesId]);
         if ($bool) {
             return response()->json(['status' => true]);
         } else {
-            return response()->json(['status' => false]);
+            return response()->json(['status' => false, 'message' => 'delete fail ...']);
         }
     }
 }
